+++ assets/js/prompt.ts
@@ -16,8 +16,8 @@ type PromptOptions = {
 
 type PromptState = 'showing' | 'endShowing' | 'hiding' | 'endHiding';
 
-function getCheckboxFromPromptContent(contentElement: HTMLElement) {
-  const root = contentElement.closest('[data-prompt]');
+function getCheckboxFromPromptContent(contentElement?: HTMLElement) {
+  const root = contentElement?.closest('[data-prompt]');
   if (!root) {
     return null; 
   }
@@ -81,31 +81,28 @@ function setCheckboxState({
   options: PromptOptions;
 }) {
   switch (state) {
-    case 'showing':
+    case "showing":
       delete checkbox.dataset.ishiding;
-      checkbox.dataset.isshowing = "";
       if (options.willShow) {
-        options.willShow(elements)
+        options.willShow(elements);
       }
       break;
-    case 'endShowing':
-      delete checkbox.dataset.isshowing;
+    case "endShowing":
       if (options.didShow) {
-        options.didShow(elements)
+        options.didShow(elements);
       }
       break;
-    case 'hiding':
-      delete checkbox.dataset.isshowing;
+    case "hiding":
       checkbox.checked = false;
-      checkbox.dataset.ishiding = "";
+      checkbox.dataset.ishiding = "true";
       if (options.willHide) {
-        options.willHide(elements)
+        options.willHide(elements);
       }
       break;
-    case 'endHiding':
+    case "endHiding":
       delete checkbox.dataset.ishiding;
       if (options.didHide) {
-        options.didHide(elements)
+        options.didHide(elements);
       }
       break;
     default:
@@ -113,8 +110,30 @@ function setCheckboxState({
   }
 }
 
-export const Prompt = {
-  mounted() {},
+type TPrompt = {
+  el?: MaybeHTMLElement;
+  checkbox?: MaybeHTMLElement;
+  init: () => void;
+  mounted: () => void;
+  updated: () => void;
+  change: (checkbox: HTMLInputElement, options: PromptOptions) => void;
+  hide: (selectorOrElement: string | HTMLElement) => void;
+  show: (selectorOrElement: string | HTMLElement) => void;
+}
+
+export const Prompt: TPrompt = {
+  init: function () {
+    this.checkbox = getCheckboxFromPromptContent(this.el || undefined);
+    if (this.checkbox) {
+      this.checkbox.dataset.ismounted = "true";
+    }
+  },
+  mounted: function () {
+    this.init();
+  },
+  updated: function () {
+    this.init();
+  },
   change: function(checkbox: HTMLInputElement, options: PromptOptions = {}) {
     const elements = getElements(checkbox);
     checkbox.addEventListener('transitionend', function(evt) {